requirements.txt




streamlit>=1.20
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0
